refactor: Database.h now exists and moved relevant functions
