+++ src/Database.h
@@ -1,0 +1,50 @@
+ #ifndef DATABASE_H
+#define DATABASE_H
+
+#include <string>
+#include <fstream>
+#include <iostream>
+#include <variant>
+#include <vector>
+
+using namespace std;
+
+vector<variant<int, string>> My_Database;
+
+void createDatabase() {
+    
+}
+
+void viewDatabase() {
+
+}
+
+void createTable() {
+    
+}
+
+void viewTable() {
+
+}
+
+void insertData() {
+
+}
+
+void viewTableCSV() {
+
+}
+
+void updateData() {
+    
+}
+
+void deleteData() {
+
+}
+
+void countRows() {
+
+}
+
+#endif
